Tried a different solution to avoid memory errors on seeding - it is not working.  The issue is that I am trying to stop the seeding every 100 records but it is difficult to restart at a given point in a very large file in Node/JS.
I tried accessing the size of the data that has
been added but using the length of the data
variable from the stream (which is the solution I
saw in several places online does not work because
my data is not parsed in the way - my data is coming
in as objects and objects do not have length) does
not work.

The easiest way would be to stop the seeding at a
certain byte size read but I cannot guarantee that
such a solution would not break across a row or
collection object.

My current attempt is having a counter for the rows
that have been inserted and comparing that number
to the number of rows that come in from the stream.
This worked fine for a much smaller dataset but is
not working correctly now.

My next step is to see if that can be refined -
otherwise I will either need to incorporate await or
try readline.